fix/refactor/change: split module into loader and parser
- split and re-organize/restructure the module consists of loader,
  dumper and other modules
- add separate test cases for each modules
